refactor: break plugin into more modules
In order to work, Quartz jobs need to be loaded at Tomcat start because they are registered with servlet mechanisms. These jobs are put into a separate `module-job` module. All classes, that are required for the job (and maybe by other plugins) are now separated into a `module-lib` module. This also has to be present at the time of Tomcat boot, to be present for other plugins and for the job during initialization. We had to duplicate the plugin name into both the plugin base and the lib to resolve cyclic dependencies. This could be resolved by better decoupling of the modules.